backend/requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
phonenumbers==8.13.45
pillow==10.4.0
python-dotenv==1.0.0
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1